+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/ProductGrid/EnrichmentSorter.php
@@ -14,12 +14,19 @@ declare(strict_types=1);
 namespace Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\ProductGrid;
 
 use Oro\Bundle\DataGridBundle\Datasource\DatasourceInterface;
+use Oro\Bundle\PimDataGridBundle\Datasource\ProductDatasource;
 use Oro\Bundle\PimDataGridBundle\Extension\Sorter\SorterInterface;
 
 final class EnrichmentSorter implements SorterInterface
 {
     public function apply(DatasourceInterface $datasource, $field, $direction)
     {
+        if (!$datasource instanceof ProductDatasource) {
+            throw new \InvalidArgumentException(
+                sprintf('Data source must be an instance of %s.', get_class(ProductDatasource::class))
+            );
+        }
+
         $datasource->getProductQueryBuilder()->addSorter($field, $direction);
     }
 }