+++ src/Akeneo/Pim/Enrichment/Bundle/Elasticsearch/Filter/Field/SelfAndAncestorFilter.php
@@ -10,7 +10,6 @@ use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\FieldFilterHelper;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\Operators;
 use Akeneo\Pim\Enrichment\Component\Product\Repository\ProductModelRepositoryInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Repository\ProductRepositoryInterface;
-use Akeneo\Tool\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 
 /**
  * An ancestor is a product model that is either a parent or a grand parent.
@@ -39,7 +38,7 @@ class SelfAndAncestorFilter extends AbstractFieldFilter
 {
     private const ANCESTOR_ID_ES_FIELD = 'ancestors.ids';
 
-    /** @var IdentifiableObjectRepositoryInterface */
+    /** @var ProductModelRepositoryInterface */
     private $productModelRepository;
 
     /** @var ProductRepositoryInterface */