+++ src/Akeneo/Pim/Enrichment/Bundle/Elasticsearch/Sorter/Field/FamilySorter.php
@@ -32,7 +32,7 @@ class FamilySorter extends BaseFieldSorter
         $this->localeRepository = $localeRepository;
     }
 
-    public function addFieldSorter($field, $direction, $locale = null, $channel = null)
+    public function addFieldSorter($field, $direction, $locale = null, $channel = null): void
     {
         if (null === $this->searchQueryBuilder) {
             throw new \LogicException('The search query builder is not initialized in the sorter.');