+++ src/Oro/Bundle/FilterBundle/Filter/FilterUtility.php
@@ -2,6 +2,8 @@
 
 namespace Oro\Bundle\FilterBundle\Filter;
 
+use Oro\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
+
 class FilterUtility
 {
     const CONDITION_OR = 'OR';
@@ -23,5 +25,14 @@ class FilterUtility
     public function getExcludeParams()
     {
         return [self::DATA_NAME_KEY, self::FORM_OPTIONS_KEY];
+    }
+
+    /**
+     * Applies filter to query by attribute
+     *
+     * @param mixed $value
+     */
+    public function applyFilter(FilterDatasourceAdapterInterface $ds, string $field, string $operator, $value)
+    {
     }
 }