+++ CHANGELOG.md
@@ -133,6 +133,15 @@
 - Remove `Akeneo\Tool\Bundle\MeasureBundle\Family\WeightFamilyInterface`
 - Rename `Akeneo\Tool\Bundle\MeasureBundle\Exception\UnknownFamilyMeasureException` as `Akeneo\Tool\Bundle\MeasureBundle\Exception\MeasurementFamilyNotFoundException`
 - Rename `Akeneo\Tool\Bundle\MeasureBundle\Exception\UnknownMeasureException` as `Akeneo\Tool\Bundle\MeasureBundle\Exception\UnitNotFoundException`
+- Change constructor of `Akeneo\Pim\Enrichment\Bundle\Command\RefreshProductCommand` to
+    - replace `Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface $productSaver` by `Akeneo\Tool\Component\StorageUtils\Saver\BulkSaverInterface $productSaver`
+    - replace `Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface $productModelSaver` by `Akeneo\Tool\Component\StorageUtils\Saver\BulkSaverInterface $productModelSaver`
+- Change constructor of `Akeneo\Pim\Enrichment\Bundle\Command\RemoveCompletenessForChannelAndLocaleCommand` to
+    - replace `Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface $channelSaver` by `Akeneo\Tool\Component\StorageUtils\Saver\BulkSaverInterface $channelSaver`
+- Interface `Akeneo\Tool\Component\Api\Repository\ApiResourceRepositoryInterface` now also extends `Doctrine\Common\Persistence\ObjectRepository`
+- Add method `getChannels()` in `Akeneo\Pim\Enrichment\Component\Category\Model\CategoryInterface` interface
+- Add method `getLabel(): string` in `Akeneo\Pim\Enrichment\Component\Category\Model\CategoryInterface` interface
+- Change `Akeneo\Pim\Enrichment\Component\Comment\Model\UserInterface` to use `Akeneo\UserManagement\Component\Model\UserInterface` instead of `Symfony\Component\Security\Core\User\UserInterface`
 
 ### CLI commands